Bandwidth selection by likelihood cross validation (multicore version)

Calculate for multiple bandiwdths the cross validation likelihood to select an appropriate bandwidth in a data driven approach. A future plan can be used to split the work across cores. The different quadras generated in accordance with the argument grid_shape are used for the parallelisation. So if only one quadra is generate (grid_shape = c(1,1)), the function will still use only one core. This also affect the progress bar.

bw_cv_likelihood_calc.mc(
  bw_range,
  bw_step,
  lines,
  events,
  w,
  kernel_name,
  method,
  diggle_correction = FALSE,
  study_area = NULL,
  max_depth = 15,
  digits = 5,
  tol = 0.1,
  agg = NULL,
  sparse = TRUE,
  grid_shape = c(1, 1),
  sub_sample = 1,
  verbose = TRUE,
  check = TRUE
)

Arguments

bw_range
The range of the bandwidths to consider, given as a numeric vector of two values: c(bandwidth_min, bandwidth_max)

bw_step
The step between each bandwidth to calculate given as a float

lines
A SpatialLinesDataFrame representing the underlying network. The geometries must be a SpatialLinesDataFrame (may crash if some geometries are invalid)

events
events A SpatialPointsDataFrame representing the events on the network. The points will be snapped on the network.

w
A vector representing the weight of each event

kernel_name
The name of the kernel to use. Must be one of triangle, gaussian, tricube, cosine ,triweight, quartic, epanechnikov or uniform.

method
The method to use when calculating the NKDE, must be one of simple / discontinuous / continuous (see details for more information)

diggle_correction
A Boolean indicating if the correction factor for edge effect must be used.

study_area
A SpatialPolygonsDataFrame or a SpatialPolygon representing the limits of the study area.

max_depth
when using the continuous and discontinuous methods, the calculation time and memory use can go wild if the network has many small edges (area with many of intersections and many events). To avoid it, it is possible to set here a maximum depth. Considering that the kernel is divided at intersections, a value of 10 should yield good estimates in most cases. A larger value can be used without problem for the discontinuous method. For the continuous method, a larger value will strongly impact calculation speed.

digits
The number of digits to retain in the spatial coordinates. It ensures that topology is good when building the network. Default is 3

tol
When adding the events and the sampling points to the network, the minimum distance between these points and the lines' extremities. When points are closer, they are added at the extremity of the lines.

agg
A double indicating if the events must be aggregated within a distance. If NULL, the events are aggregated by rounding the coordinates.

sparse
A Boolean indicating if sparse or regular matrix should be used by the Rcpp functions. Regular matrix are faster, but require more memory and could lead to error, in particular with multiprocessing. Sparse matrix are slower, but require much less memory (not used for the moment).

grid_shape
A vector of two values indicating how the study area must be split when performing the calculus (see details). Default is c(1,1)

sub_sample
A float between 0 and 1 indicating the percentage of quadra to keep in the calculus. For large datasets, it may be useful to limit the bandwidth evaluation and thus reduce calculation time.

verbose
A Boolean, indicating if the function should print messages about process.

check
A Boolean indicating if the geometry checks must be run before calculating the density.

Value

A dataframe with two columns, one for the bandwidths and the second for the cross validation score (the lower the better).

Details

For more details, see help(bw_cv_likelihood_calc)

Examples

# \donttest{
networkgpkg <- system.file("extdata", "networks.gpkg", package = "spNetwork", mustWork = TRUE)
eventsgpkg <- system.file("extdata", "events.gpkg", package = "spNetwork", mustWork = TRUE)
mtl_network <- rgdal::readOGR(networkgpkg,layer="mtl_network", verbose=FALSE)
bike_accidents <- rgdal::readOGR(eventsgpkg,layer="bike_accidents", verbose=FALSE)
future::plan(future::multisession(workers=2))
cv_scores <- bw_cv_likelihood_calc.mc(c(200,800),50,
                               mtl_network, bike_accidents,
                               rep(1,nrow(bike_accidents)),
                               "quartic", "simple",
                               diggle_correction = FALSE, study_area = NULL,
                               max_depth = 8,
                               digits=2, tol=0.1, agg=5,
                               sparse=TRUE, grid_shape=c(1,1),
                               sub_sample = 1, verbose=TRUE, check=TRUE)

#> [1] "checking inputs ..."
#> [1] "prior data preparation ..."
#> [1] "Calculating the correction factor if required"

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Warning: spgeom1 and spgeom2 have different proj4 strings

#> Warning: CRS object has comment, which is lost in output

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> Warning: number of rows of result is not a multiple of vector length (arg 2)

#> [1] "start calculating the CV values ..."

#> Warning: UNRELIABLE VALUE: One of the ‘future.apply’ iterations (‘future_lapply-1’) unexpectedly generated random numbers without declaring so. There is a risk that those random numbers are not statistically sound and the overall results might be invalid. To fix this, specify 'future.seed=TRUE'. This ensures that proper, parallel-safe random numbers are produced via the L'Ecuyer-CMRG method. To disable this check, use 'future.seed = NULL', or set option 'future.rng.onMisuse' to "ignore".

## make sure any open connections are closed afterward
if (!inherits(future::plan(), "sequential")) future::plan(future::sequential)
# }