Time and Network bandwidth correction calculation for arrays

Calculating the border correction factor for both time and network bandwidths when we have to deal with adaptive bandwidths and arrays

bw_tnkde_corr_factor_arr(
  net_bws,
  time_bws,
  diggle_correction,
  study_area,
  events,
  events_loc,
  lines,
  method,
  kernel_name,
  tol,
  digits,
  max_depth,
  sparse,
  time_limits = NULL
)

Arguments

net_bws

An array of network bandwidths

time_bws

An array of time bandwidths

diggle_correction

A Boolean indicating if the correction factor for edge effect must be used.

study_area

A feature collection of polygons representing the limits of the study area.

events

A feature collection of points representing the events

events_loc

A feature collection of points representing the unique location of events

lines

A feature collection of linestrings representing the underlying lines of the network

method

The name of the NKDE to use

kernel_name

The name of the kernel to use

tol

float indicating the minimum distance between the events and the lines' extremities when adding the point to the network. When points are closer, they are added at the extremity of the lines.

digits

An integer, the number of digits to keep for the spatial coordinates

max_depth

The maximal depth for continuous or discontinuous NKDE

sparse

A Boolean indicating if sparse or regular matrices should be used by the Rcpp functions. These matrices are used to store edge indices between two nodes in a graph. Regular matrices are faster, but require more memory, in particular with multiprocessing. Sparse matrices are slower (a bit), but require much less memory.

time_limits

A vector with the upper and lower limit of the time period studied

Examples

# no example provided, this is an internal function