The main function to calculate continuous NKDE (with ARMADILO and sparse matrix)

The main function to calculate continuous NKDE (with ARMADILO and sparse matrix)

The main function to calculate continuous NKDE (with ARMADILO and integer matrix)

continuous_nkde_cpp_arma_sparse(
  neighbour_list,
  events,
  weights,
  samples,
  bws,
  kernel_name,
  nodes,
  line_list,
  max_depth,
  verbose,
  div = "bw"
)

continuous_nkde_cpp_arma(
  neighbour_list,
  events,
  weights,
  samples,
  bws,
  kernel_name,
  nodes,
  line_list,
  max_depth,
  verbose,
  div = "bw"
)

Arguments

neighbour_list

a list of the neighbours of each node

events

a numeric vector of the node id of each event

weights

a numeric vector of the weight of each event

samples

a DataFrame of the samples (with spatial coordinates and belonging edge)

bws

the kernel bandwidths for each event

kernel_name

the name of the kernel to use

nodes

a DataFrame representing the nodes of the graph (with spatial coordinates)

line_list

a DataFrame representing the lines of the graph

max_depth

the maximum recursion depth (after which recursion is stopped)

verbose

a boolean indicating if the function must print its progress

div

The divisor to use for the kernel. Must be "n" (the number of events within the radius around each sampling point), "bw" (the bandwidth) "none" (the simple sum).

Value

a DataFrame with two columns : the kernel values (sum_k) and the number of events for each sample (n) a DataFrame with two columns : the kernel values (sum_k) and the number of events for each sample (n)