c++ k and g function counting worker

c++ k function counting (INTERNAL)

c++ k function counting (INTERNAL)

kgfunc_time_counting(
  dist_mat_net,
  dist_mat_time,
  wc,
  wr,
  breaks_net,
  breaks_time,
  width_net,
  width_time,
  cross = FALSE
)

kfunc_time_counting(
  dist_mat_net,
  dist_mat_time,
  wc,
  wr,
  breaks_net,
  breaks_time,
  cross = FALSE
)

Arguments

dist_mat_net

A matrix with the distances between points on the network

dist_mat_time

A matrix with the distances between points in time

wc

The weight of the points represented by the columns (destinations)

wr

The weight of the points represented by the rows (origins)

breaks_net

A numeric vector with the distance to consider on network

breaks_time

A numeric vector with the distance to consider in time

width_net

The width of each donut for the network dimension

width_time

The width of each donut for the time dimension

cross

A boolean indicating if we are calculating a cross k function or not (default is FALSE)

Value

A list of two numeric cubes with the values of the k and g function evaluated at the required distances

A list of two numeric cubes with the values of the k and g function evaluated at the required distances