network_listw worker

The worker function of network_listw.

network_listw_worker(
  points,
  lines,
  maxdistance,
  dist_func,
  direction = NULL,
  mindist = 10,
  matrice_type = "B",
  verbose = FALSE,
  digits = 3,
  tol = 0.1
)

Arguments

points

A feature collection of points corresponding to start and end points. It must have a column fid, grouping the points if necessary.

lines

A feature collection of lines representing the network

maxdistance

The maximum distance between two observation to consider them as neighbours.

dist_func

A vectorized function converting spatial distances into weights.

direction

Indicate a field giving information about authorized travelling direction on lines. if NULL, then all lines can be used in both directions. Must be the name of a column otherwise. The values of the column must be "FT" (From - To), "TF" (To - From) or "Both".

mindist

The minimum distance between two different observations. It is important for it to be different from 0 when a W style is used.

matrice_type

The type of the weighting scheme. Can be 'B' for Binary, 'W' for row weighted, or 'I' (identity), see the documentation of spdep::nb2listw for details

verbose

A Boolean indicating if the function should print its progress

digits

the number of digits to keep in the spatial coordinates ( simplification used to reduce risk of topological error)

tol

A float indicating the spatial tolerance when points are added as vertices to lines.

Value

A list of neihbours as weights.

Examples

#no example provided, this is an internal function